Rename: ByteBlockGenerator to RSAKeyGenerator
